libkvs/classic: reimplement kvsdir_get_*()
