+++ src/common/libkvs/kvs.c
@@ -49,7 +49,6 @@
 #include "src/common/libutil/xzmalloc.h"
 #include "src/common/libutil/blobref.h"
 
-#include "kvs_deprecated.h"
 #include "proto.h"
 #include "json_dirent.h"
 
@@ -158,55 +157,6 @@ done:
     Jput (rootref_obj);
     flux_future_destroy (f);
     errno = saved_errno;
-    return rc;
-}
-
-int kvs_get_dirat (flux_t *h, const char *rootref,
-                   const char *key, kvsdir_t **dir)
-{
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (!rootref || !key || !dir) {
-        errno = EINVAL;
-        goto done;
-    }
-    if (getobj (h, rootref, key, KVS_PROTO_READDIR, &v) < 0)
-        goto done;
-    *dir = kvsdir_create (h, rootref, key, Jtostr (v));
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
-
-int kvs_get_dir (flux_t *h, kvsdir_t **dir, const char *fmt, ...)
-{
-    va_list ap;
-    char *key = NULL;
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (!h || !dir || !fmt) {
-        errno = EINVAL;
-        goto done;
-    }
-    va_start (ap, fmt);
-    key = xvasprintf (fmt, ap);
-    va_end (ap);
-
-    /* N.B. python kvs tests use empty string key for some reason.
-     * Don't break them for now.
-     */
-    const char *k = strlen (key) > 0 ? key : ".";
-    if (getobj (h, NULL, k, KVS_PROTO_READDIR, &v) < 0)
-        goto done;
-    *dir = kvsdir_create (h, NULL, k, Jtostr (v));
-    rc = 0;
-done:
-    Jput (v);
-    if (key)
-        free (key);
     return rc;
 }
 
@@ -1018,170 +968,6 @@ done:
 /**
  ** kvsdir_t convenience functions
  **/
-
-static int dirgetobj (kvsdir_t *dir, const char *name,
-                      int flags, json_object **val)
-{
-    flux_t *h = kvsdir_handle (dir);
-    char *key;
-    int rc;
-
-    key = kvsdir_key_at (dir, name);
-    rc = getobj (h, kvsdir_rootref (dir), key, flags, val);
-    free (key);
-    return rc;
-}
-
-int kvsdir_get (kvsdir_t *dir, const char *name, char **valp)
-{
-    json_object *v = NULL;
-    if (dirgetobj (dir, name, 0, &v) < 0)
-        return -1;
-    if (valp)
-        *valp = xstrdup (Jtostr (v));
-    Jput (v);
-    return 0;
-}
-
-int kvsdir_get_dir (kvsdir_t *dir, kvsdir_t **dirp, const char *fmt, ...)
-{
-    flux_t *h = kvsdir_handle (dir);
-    int rc = -1;
-    char *name, *key;
-    va_list ap;
-    json_object *v = NULL;
-
-    va_start (ap, fmt);
-    if (vasprintf (&name, fmt, ap) < 0)
-        oom ();
-    va_end (ap);
-
-    key = kvsdir_key_at (dir, name);
-    if (getobj (h, kvsdir_rootref (dir), key, KVS_PROTO_READDIR, &v) < 0)
-        goto done;
-    *dirp = kvsdir_create (h, kvsdir_rootref (dir), key, Jtostr (v));
-    rc = 0;
-done:
-    Jput (v);
-    free (key);
-    free (name);
-    return rc;
-}
-
-int kvsdir_get_symlink (kvsdir_t *dir, const char *name, char **valp)
-{
-    int rc = -1;
-    json_object *v = NULL;
-
-    if (dirgetobj (dir, name, KVS_PROTO_READLINK, &v) < 0)
-        goto done;
-    if (json_object_get_type (v) != json_type_string) {
-        errno = EPROTO;
-        goto done;
-    }
-    if (valp)
-        *valp = xstrdup (json_object_get_string (v));
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
-
-int kvsdir_get_string (kvsdir_t *dir, const char *name, char **valp)
-{
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (dirgetobj (dir, name, 0, &v) < 0)
-        goto done;
-    if (json_object_get_type (v) != json_type_string) {
-        errno = EPROTO;
-        goto done;
-    }
-    if (valp)
-        *valp = xstrdup (json_object_get_string (v));
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
-
-int kvsdir_get_int (kvsdir_t *dir, const char *name, int *valp)
-{
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (dirgetobj (dir, name, 0, &v) < 0)
-        goto done;
-    if (json_object_get_type (v) != json_type_int) {
-        errno = EPROTO;
-        goto done;
-    }
-    if (valp)
-        *valp = json_object_get_int (v);
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
-
-int kvsdir_get_int64 (kvsdir_t *dir, const char *name, int64_t *valp)
-{
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (dirgetobj (dir, name, 0, &v) < 0)
-        goto done;
-    if (json_object_get_type (v) != json_type_int) {
-        errno = EPROTO;
-        goto done;
-    }
-    if (valp)
-        *valp = json_object_get_int64 (v);
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
-
-int kvsdir_get_double (kvsdir_t *dir, const char *name, double *valp)
-{
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (dirgetobj (dir, name, 0, &v) < 0)
-        goto done;
-    if (json_object_get_type (v) != json_type_double
-        && json_object_get_type (v) != json_type_int) {
-        errno = EPROTO;
-        goto done;
-    }
-    if (valp)
-        *valp = json_object_get_double (v);
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
-
-int kvsdir_get_boolean (kvsdir_t *dir, const char *name, bool *valp)
-{
-    json_object *v = NULL;
-    int rc = -1;
-
-    if (dirgetobj (dir, name, 0, &v) < 0)
-        goto done;
-    if (json_object_get_type (v) != json_type_boolean) {
-        errno = EPROTO;
-        goto done;
-    }
-    if (valp)
-        *valp = json_object_get_boolean (v);
-    rc = 0;
-done:
-    Jput (v);
-    return rc;
-}
 
 int kvsdir_put (kvsdir_t *dir, const char *name, const char *val)
 {
